// frontend/src/api/schema.d.ts
export interface components {
    schemas: {
        /**
         * Category
         * @enum {string}
         */
        Category: "water" | "electricity" | "sanitation" | "roads" | "streetlights" | "other";
// ...
        Priority: "high" | "normal" | "low";
// ...
        Status: "open" | "in_progress" | "resolved" | "rejected";
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}